Move gfx unit tests to the unit_tests crate.
