+++ tests/unit/Cargo.toml
@@ -11,5 +11,8 @@ doctest = false
 [dependencies.util]
 path = "../../components/util"
 
+[dependencies.gfx]
+path = "../../components/gfx"
+
 [dependencies.geom]
 git = "https://github.com/servo/rust-geom"
