+++ tests/unit/lib.rs
@@ -1,4 +1,6 @@
 extern crate geom;
+extern crate gfx;
 extern crate util;
 
 #[cfg(test)] #[path="util/mod.rs"] mod util_tests;
+#[cfg(test)] #[path="gfx/mod.rs"] mod gfx_tests;
